pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sql-debug-env"
version = "0.1.0"
description = "OpenEnv environment for SQL debugging and optimisation"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.2",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "websockets>=11.0",
    "httpx>=0.25.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.setuptools]
package-dir = {"sql_debug_env" = ".", "sql_debug_env.server" = "server"}
packages = ["sql_debug_env", "sql_debug_env.server", "sql_debug_env.server.tasks"]

[project.scripts]
server = "sql_debug_env.server.app:main"
